+++ api/src/index.js
@@ -26,48 +26,3 @@ controller.hears(['list'], ['direct_message'], () => {
     data.list()
 })
 
-controller.hears(
-    ['Con tên gì thế?', 'chào con'],
-    ['direct_message', 'direct_mention', 'mention', 'ambient'],
-    (bot, message) => {
-        bot.api.users.info({ user: message.user }, (error, response) => {
-            const { display_name } = response.user.profile
-            bot.reply(
-                message,
-                `dạ con chào cô/cậu ${display_name} con tên là shin ạ!`,
-            )
-        })
-    },
-)
-
-controller.hears(
-    ['Con tên gì thế?', 'chào con'],
-    ['direct_message', 'direct_mention', 'mention', 'ambient'],
-    (bot, message) => {
-        bot.api.users.info({ user: message.user }, (error, response) => {
-            const { display_name } = response.user.profile
-
-            if (display_name === 'shun') {
-                bot.reply(
-                    message,
-                    `dạ con chào cậu ${display_name} con tên là shin ạ!`,
-                )
-            } else {
-                bot.reply(
-                    message,
-                    `dạ con chào cô ${display_name} con tên là shin ạ!`,
-                )
-            }
-        })
-    },
-)
-
-controller.hears(
-    ['con thấy cô quyên xinh hem?'],
-    ['direct_message', 'direct_mention', 'mention'],
-    (bot, message) => {
-        controller.storage.users.get(message.user, () => {
-            bot.reply(message, 'Meow ! cô quyên xinh nhất Quảng Yên Meow!')
-        })
-    },
-)
