+++ api/src/util/index.js
@@ -1,0 +1,20 @@
+module.exports.ConvertRowsToObject = (rows) => {
+    const results = []
+
+    if (!rowss || rows.length === 0) {
+        return []
+    }
+
+    const headers = rows[0]
+
+    rows.map((values, index) => {
+        if (index !== 0) {
+            results[index - 1] = {}
+            headers.map((value, i) => {
+                results[index - 1][value] = values[i]
+            })
+        }
+    })
+
+    return results
+}
